+++ platform/flowglad-next/llm-prompts/new-combined-db-table-protected-procedures-and-form-modals.txt
@@ -265,7 +265,7 @@ import {
 } from '@/db/schema/unicornRiders'
 
 import { protectedProcedure } from '@/server/trpc'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import {
   insertUnicornRider,
   selectUnicornRidersPaginated,
