+++ platform/flowglad-next/src/app/onboarding/page.tsx
@@ -1,5 +1,5 @@
 import { adminTransaction } from '@/db/adminTransaction'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import { selectMembershipAndOrganizations } from '@/db/tableMethods/membershipMethods'
 import {
   FlowgladApiKeyType,
