+++ platform/flowglad-next/src/app/store/products/page.tsx
@@ -1,6 +1,6 @@
 import * as R from 'ramda'
 import Internal from './Internal'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import { Price } from '@/db/schema/prices'
 import { selectMembershipAndOrganizations } from '@/db/tableMethods/membershipMethods'
 import {
