+++ platform/flowglad-next/src/server/mutations/createFile.ts
@@ -1,6 +1,6 @@
 import { protectedProcedure } from '@/server/trpc'
 import { createFileInputSchema } from '@/db/schema/files'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import { insertFileTransaction } from '@/utils/fileStorage'
 
 export const createFile = protectedProcedure
