+++ platform/flowglad-next/src/server/mutations/deleteLink.ts
@@ -1,5 +1,5 @@
 import { protectedProcedure } from '@/server/trpc'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import { idInputSchema } from '@/db/tableUtils'
 import { deleteLink } from '@/db/tableMethods/linkMethods'
