+++ platform/flowglad-next/src/server/mutations/toggleTestMode.ts
@@ -1,6 +1,6 @@
 import { protectedProcedure } from '@/server/trpc'
 import { adminTransaction } from '@/db/adminTransaction'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import {
   selectFocusedMembershipAndOrganization,
   updateMembership,
