+++ platform/flowglad-next/src/server/queries/getCustomer.ts
@@ -1,5 +1,5 @@
 import { protectedProcedure } from '@/server/trpc'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import { z } from 'zod'
 import { selectCustomers } from '@/db/tableMethods/customerMethods'
 import { customerClientSelectSchema } from '@/db/schema/customers'
