+++ platform/flowglad-next/src/server/routers/checkoutSessionsRouter.ts
@@ -4,7 +4,7 @@ import {
   publicProcedure,
   router,
 } from '@/server/trpc'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import {
   insertCheckoutSession,
   selectCheckoutSessionById,
