+++ platform/flowglad-next/src/server/routers/discountsRouter.ts
@@ -16,7 +16,7 @@ import {
 } from '@/db/schema/discounts'
 
 import { protectedProcedure } from '@/server/trpc'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import { createDiscountInputSchema } from '@/db/schema/discounts'
 import {
   insertDiscount,
