+++ platform/flowglad-next/src/server/routers/paymentMethodsRouter.ts
@@ -1,5 +1,5 @@
 import { protectedProcedure, router } from '@/server/trpc'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import {
   selectPaymentMethodById,
   selectPaymentMethodsPaginated,
