+++ platform/flowglad-next/src/server/routers/pricesRouter.ts
@@ -1,6 +1,6 @@
 import { router } from '@/server/trpc'
 import { protectedProcedure } from '@/server/trpc'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import {
   editPriceSchema,
   pricesClientSelectSchema,
