+++ platform/flowglad-next/src/server/routers/productsRouter.ts
@@ -17,7 +17,7 @@ import {
   editProductSchema,
   productWithPricesSchema,
 } from '@/db/schema/prices'
-import { authenticatedTransaction } from '@/db/databaseMethods'
+import { authenticatedTransaction } from '@/db/authenticatedTransaction'
 import { selectMembershipAndOrganizations } from '@/db/tableMethods/membershipMethods'
 import { generateOpenApiMetas, trpcToRest } from '@/utils/openapi'
 import { z } from 'zod'
